requirements.txt
requests
yfinance
vaderSentiment
python-dotenv
streamlit
pandas
